Rename spinner package to progress
