+++ internal/pkg/term/progress/mocks/mock_spinner.go
