+++ internal/pkg/term/progress/spinner.go
@@ -1,8 +1,8 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-// Package spinner provides a simple wrapper around github.combriandowns/spinner to start and stop a spinner in the terminal.
-package spinner
+// Package progress provides a simple wrapper around github.combriandowns/spinner to start and stop a spinner in the terminal.
+package progress
 
 import (
 	"fmt"
@@ -54,8 +54,8 @@ type Spinner struct {
 	eventsWriter writeFlusher // Writer to pretty format events in a table.
 }
 
-// New returns a spinner that outputs to stderr.
-func New() *Spinner {
+// NewSpinner returns a spinner that outputs to stderr.
+func NewSpinner() *Spinner {
 	s := spinner.New(charset, 125*time.Millisecond, spinner.WithHiddenCursor(true))
 	s.Writer = os.Stderr
 	return &Spinner{
